+++ src/Simplyhired.php
@@ -178,16 +178,6 @@ class Simplyhired extends AbstractProvider
     }
 
     /**
-     * Get parameters
-     *
-     * @return  array
-     */
-    public function getParameters()
-    {
-        return [];
-    }
-
-    /**
      * Get query string for client based on properties
      *
      * @return string